assignment_7_kafka_streaming/requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
confluent-kafka==1.9.2

oauthlib==3.2.1
pygogo==1.3.0
python-dotenv==0.21.0
requests==2.28.1
requests-oauthlib==1.3.1
tweepy==4.10.1
urllib3==1.26.12
